requirements.txt
auth0-python==3.1.2
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
gunicorn==19.7.1
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
psycopg2==2.7.3
python-dotenv==0.6.5
requests==2.18.3
SQLAlchemy==1.1.13
urllib3==1.22
Werkzeug==0.12.2
WTForms==2.1
